+++ src/main/java/fpt/capstone/betatest/services/KeywordService.java
@@ -59,7 +59,7 @@ public class KeywordService {
 	}
 
 	@Transactional
-	public MessageOutputModel createKeyword(User user, Keyword kw, String keyword, float crisisRate) {
+	public MessageOutputModel createKeyword(User user, Keyword kw, String keyword, double crisisRate) {
 		boolean havePermissionToCreate = false;
 		MessageOutputModel mod = new MessageOutputModel();
 		if (!(user.getRole().equals("user") && !user.isAvailable())) {
@@ -107,7 +107,7 @@ public class KeywordService {
 	}
 
 	@Transactional
-	public MessageOutputModel updateKeyword(User user, Keyword kw, String keyword, int log_version, int keywordId, float crisisRate) {
+	public MessageOutputModel updateKeyword(User user, Keyword kw, String keyword, int log_version, int keywordId, double crisisRate) {
 		boolean havePermissionToUpdate = false;
 		MessageOutputModel mod = new MessageOutputModel();
 		if ((user.getUserName().equals(kw.getUser().getUserName())) || user.getRole().equals("admin")) {
